+++ src/components/Layout/IfFeatureEnable/index.ts
@@ -1,4 +1,0 @@
-import IfFeatureEnable from "./IfFeatureEnable";
-
-export default IfFeatureEnable;
-export * from "./IfFeatureEnable";