+++ src/components/Layout/IfFeatureEnable/stories/IfFeatureEnable.tsx
@@ -1,5 +1,0 @@
-import IfFeatureEnableComponent, { IfFeatureEnableProps } from "@/components/Layout/IfFeatureEnable/IfFeatureEnable";
-
-const IfFeatureEnable = (props: IfFeatureEnableProps) => <IfFeatureEnableComponent {...props} />;
-
-export default IfFeatureEnable;