+++ src/components/Utils/HasPermission/index.ts
@@ -1,0 +1,4 @@
+import HasPermission from "./HasPermission";
+
+export default HasPermission;
+export * from "./HasPermission";