+++ src/components/Utils/HasPermission/test/HasPermission.test.tsx
@@ -1,13 +1,13 @@
 import { render } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
-import IfFeatureEnable from "../IfFeatureEnable";
+import HasPermission from "../HasPermission";
 
-describe("Test <IfFeatureEnable/>", () => {
+describe("Test <HasPermission/>", () => {
   it("with string name", () => {
     const { getByText } = render(
-      <IfFeatureEnable name="featureName" features={["featureName"]}>
+      <HasPermission name="featureName" additionalPermissions={["featureName"]}>
         This should render
-      </IfFeatureEnable>,
+      </HasPermission>,
     );
 
     expect(getByText("This should render")).toBeInTheDocument();
@@ -15,9 +15,9 @@ describe("Test <IfFeatureEnable/>", () => {
 
   it("with string name and several features", () => {
     const { getByText } = render(
-      <IfFeatureEnable name="featureName" features={["featureName", "featureName2"]}>
+      <HasPermission name="featureName" additionalPermissions={["featureName", "featureName2"]}>
         This should render
-      </IfFeatureEnable>,
+      </HasPermission>,
     );
 
     expect(getByText("This should render")).toBeInTheDocument();
@@ -25,9 +25,9 @@ describe("Test <IfFeatureEnable/>", () => {
 
   it("with feature name not allowed", () => {
     const { queryByText } = render(
-      <IfFeatureEnable name="featureName1" features={["featureName2"]}>
+      <HasPermission name="featureName1" additionalPermissions={["featureName2"]}>
         This not should render
-      </IfFeatureEnable>,
+      </HasPermission>,
     );
 
     expect(queryByText("This not should render")).not.toBeInTheDocument();
@@ -35,9 +35,9 @@ describe("Test <IfFeatureEnable/>", () => {
 
   it("with array of feature name", () => {
     const { getByText } = render(
-      <IfFeatureEnable name={["featureName1", "featureName2"]} features={["featureName1", "featureName2", "featureName3"]}>
+      <HasPermission name={["featureName1", "featureName2"]} additionalPermissions={["featureName1", "featureName2", "featureName3"]}>
         This should render
-      </IfFeatureEnable>,
+      </HasPermission>,
     );
 
     expect(getByText("This should render")).toBeInTheDocument();
@@ -45,9 +45,9 @@ describe("Test <IfFeatureEnable/>", () => {
 
   it("with array of feature name with some missing name", () => {
     const { getByText } = render(
-      <IfFeatureEnable name={["featureName1", "featureName4"]} features={["featureName1", "featureName2", "featureName3"]}>
+      <HasPermission name={["featureName1", "featureName4"]} additionalPermissions={["featureName1", "featureName2", "featureName3"]}>
         This should render
-      </IfFeatureEnable>,
+      </HasPermission>,
     );
 
     expect(getByText("This should render")).toBeInTheDocument();
@@ -55,9 +55,9 @@ describe("Test <IfFeatureEnable/>", () => {
 
   it("with array with missing name", () => {
     const { queryByText } = render(
-      <IfFeatureEnable name={["featureName4", "featureName5"]} features={["featureName1", "featureName2", "featureName3"]}>
+      <HasPermission name={["featureName4", "featureName5"]} additionalPermissions={["featureName1", "featureName2", "featureName3"]}>
         This not should render
-      </IfFeatureEnable>,
+      </HasPermission>,
     );
 
     expect(queryByText("This not should render")).not.toBeInTheDocument();