+++ src/hooks/useFeatureEnabled/index.ts
@@ -1,4 +1,0 @@
-import useFeatureEnabled from "./useFeatureEnabled";
-
-export default useFeatureEnabled;
-export * from "./useFeatureEnabled";