+++ src/hooks/usePermission/index.ts
@@ -1,0 +1,4 @@
+import usePermission from "./usePermission";
+
+export default usePermission;
+export * from "./usePermission";