+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v2.0.1
-- **[fix]** - AppBar : center vertical logo
+## v2.0.2
+- **[refactor]** - rename hook `useFeatureEnabled` => `usePermission`
+- **[refactor]** - rename component `IfFeatureEnable` => `HasPermission`
+- **[refactor]** - rename provider `FeatureEnableProvider` => `PermissionProvider`
+- **[refactor]** - rename function `getIsFeatureEnabled` => `hasPermission`
+- **[refactor]** - rename function `appendFeature` => `appendPermissions`